+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "converse.js",
-  "version": "7.0.2",
+  "version": "8.0.0",
   "description": "Browser based XMPP chat client",
   "browser": "dist/converse.js",
   "module": "src/converse.js",
