+++ src/headless/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@converse/headless",
-  "version": "7.0.2",
+  "version": "8.0.0",
   "description": "Converse.js Headless build",
   "author": "cmrd Senya <senya@riseup.net>",
   "homepage": "https://conversejs.org",
