+++ src/plugins/chatboxviews/index.js
@@ -8,14 +8,10 @@ import '@converse/headless/plugins/chatboxes/index.js';
 import ChatBoxViews from './container.js';
 import { ViewWithAvatar } from 'shared/avatar.js';
 import { _converse, api, converse } from '@converse/headless/core';
+import { calculateViewportHeightUnit } from './utils.js';
 
 import './styles/chats.scss';
 
-
-function calculateViewportHeightUnit () {
-    const vh = window.innerHeight * 0.01;
-    document.documentElement.style.setProperty('--vh', `${vh}px`);
-}
 
 converse.plugins.add('converse-chatboxviews', {
     dependencies: ['converse-chatboxes', 'converse-vcard'],
