+++ src/plugins/chatboxviews/utils.js
@@ -1,0 +1,5 @@
+
+export function calculateViewportHeightUnit () {
+    const vh = window.innerHeight * 0.01;
+    document.documentElement.style.setProperty('--vh', `${vh}px`);
+}
