+++ src/plugins/chatview/templates/toolbar.js
@@ -1,5 +1,6 @@
-import { html } from "lit-html";
+import 'shared/chat/toolbar.js';
 import { api } from '@converse/headless/core.js';
+import { html } from "lit-html";
 
 export default (o) => {
     const message_limit = api.settings.get('message_limit');
