+++ src/plugins/muc-views/bottom-panel.js
@@ -1,3 +1,4 @@
+import 'shared/autocomplete/index.js';
 import BottomPanel from 'plugins/chatview/bottom-panel.js';
 import debounce from 'lodash/debounce';
 import tpl_muc_bottom_panel from './templates/muc-bottom-panel.js';
