+++ src/shared/chat/templates/new-day.js
